New API to unlock additional performance
What performance can I expect from this change? As an anecdote, I've
been testing this update on the rocket league replay parser,
[boxcars][boxcars], and saw the entire program throughput increase by
20%. Your mileage may vary depending on if one can take advantage of
patterns while decoding.

MSRV bumped to 1.48.0

I want to first apologize to downstream users. The API changes are quite
extensive, but hopefully by the end you will agree the changes are worth
it.

The most dramatic change is that native bit reads are now limited to 56
bits. Larger bit reads now need to be emulated. Even these emulated
reads are still faster than other libraries. It is now a debug time
assertion if a read size exceeds 56 bits.

Bitter is substantially slimmer with this update. Most of the
`unchecked` APIs have been removed. APIs like `if_get`, `read_max_bits`,
`bit_width` have been removed, as these can be implemented in a more
performant manner where patterns in downstream data can be capitalized
on. `approx_bytes_remaining` has been renamed to `bytes_remaining` and
the behavior updated to mean the number of complete bytes left to read
in the buffer.

The `i32` bit size input arguments have been changed to `u32`, which is
more appropriate (can't be less than zero) and didn't come with a
performance hit that `u8` and `usize` came with.

What spurred this change was a [wonderful article][dougallj] about how
optimizing a bit reader around [Fabian Giesen's articles][fgiesen] would
lead to large gains. This got me thinking how these could be translated
to bitter, as I often find bit reading to be a significant hotspot in
projects of mine. So I got to work. Basically the new API allows
lookahead buffer refilling and consumption to be controlled by the user.

Failed Attempts
---------------

In an attempt to be clever by eliminating end of buffer checks, I tried
creating an API where one would give a buffer and a length that would be
at least 8 less than the length of the buffer. That way, all refills
could be an unaligned read without a bound checks.

Well, there still needs to be a bounds check somewhere to inform the
downstream user that there are fewer bits to read, and that bounds check
cut the performance of this experiment (called `xl` for extra long
buffer) in half, such that it was slower than the manual api.

Then I figured if one was bent on removing bound checks then they can
use the unchecked APIs

[dougallj]: https://dougallj.wordpress.com/2022/08/20/faster-zlib-deflate-decompression-on-the-apple-m1-and-x86
[fgiesen]: https://fgiesen.wordpress.com/2018/02/20/reading-bits-in-far-too-many-ways-part-2
[boxcars]: https://github.com/nickbabcock/boxcars