Resolving remaining pedantic clippy issues
The only change to public APIs is the annotation of `#[must_use]`, but
I'm not considering this a breaking change as this would most likely
reveal the presence of a bug.

I verified that the assembly did not change with the read_eof
implementation (using `#[inline(never)]`). Amusingly, whitespace
changed. I'm not sure why, but it doesn't seem like a problem.