Replace macro generated bit readers with const generic (#34)
Having one giant macro define everything wasn't great for DX, and so I'm
replacing it with const generic.

To support const generics, MSRV is bumped from 1.48 to 1.51. This can be
considered a breaking change I suppose, though I won't be considering it as
such as 1.51 was released nearly 3 years ago, which is much much longer than
the time between 1.48 and 1.51

Most of the code additions are due to simply forwarding the trait
implementation to the underlying implementation. Originally, I transitioned
the little and big readers to be type aliases but that would require exposing
the internals and would break semver as a type alias is not a substitution
for a struct (at least from cargo semver check).

So I went with preserving backwards compatibility at the cost of 200 lines of
trivial trait forwarding. Not enough to warrant looking at trait delegation
traits like `delegate` and `ambassador`