Support bit reads of size zero
When reading a length prefixed value, that length could be zero. The
current behavior when reading zero is to panic as the big endian reader
would witness a right shift overflow.

This commit removes the restriction at the cost of an extra mask for big
endian readers. This causes up to a 10% throughput penalty (with
averages in the low single digits) for the big endian reader. This
downside seems acceptable.