Reduce auto mode refills
Previously `read_bits` required the lookahead to contain all the data so
that only a single peek is required.

This commit changes this by consuming the entire lookahead and shifting
in more data.

This reduces the number of refills (and thus memory loads). The exact
reduction eludes me, but previously reading 32 bits at a bit offset
would average one read per refill, and now the number is much closer to
2 (so upto a 2x reduction).