Relax peek/consume constraints with exposure of remaining lookahead bits
This is a breaking change, but one that is hopefully understood.

Previously one would be unable to `peek` or `consume` more than
`MAX_READ_BITS` at once even though it is possible for a given refill to
support more than `MAX_READ_BITS`. This commit removes the constraint so
that only the amount of bits in the lookahead buffer is the max for a
given peek or consume. (This also could be worked around by compiling
without debug assertions, but we don't talk about that).

Since the number of available bits in the lookahead buffer is already
represented by the internal state `bit_count`, it made sense to drop the
return value from `refill_lookahead` and add a new function
`lookahead_bits` in case one wanted to know the number of available bits
at any point in time.

Benchmarks have been updated to use this new technique.