Remove black_box in benchmarks in favor of accumulator
When benchmarking bitter changes at macro level, I noticed it was not
reflective of the benchmarking code within bitter. I believe a big part
of that reason is due to the use of `black_box` preventing optimizations
that are apparent in any practical usage.

An accumulator + assertion replaces `black_box` which should allow for
all bit streams to stretch their proverbial legs.